pyproject.toml
[project]
name = "entrezweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "biopython>=1.85",
  "celery[redis]>=5.5.3",
  "django>=5.2.4",
  "django-celery-results>=2.6.0",
  "django-compressor>=4.5.1",
  "django-debug-toolbar>=5.2.0",
  "django-extensions>=4.1",
  "django-htmx>=1.23.2",
  "django-ninja>=1.4.3",
  "django-render-block>=0.11",
  "flower>=2.0.1",
  "psycopg2-binary>=2.9.10",
  "pydantic>=2.11.7",
  "python-dotenv>=1.1.1",
  "requests>=2.32.4",
]

[dependency-groups]
dev = ["ipdb>=0.13.13", "ipython>=9.4.0"]

[project.scripts]
"manage.py" = "eweb:django_manage"
